Create run.py
This is my initial commit. Please note that this is just a hobby project and I am not a python programmer. I am sure there are still bugs in there. Use it at your own risk to implement your needs :) 

This script allows to

- Limit the maximum charge current of home battery
- Limit the maximum discharge current of home battery
- Set a new Grid Point
- Toggle Single/Three Phase

dbus python libraries are used to realize these settings.

The goal is to modify how the Fronius Wattpilot with its default setting is working.

Eco Mode: 
- Limit MP2 discharge current
- Apply MP2 dynamic charge profile (charge load depends on SoC of home battery)

The goal is to use this setting for load our EV car with maximum affort. The discharge is 1380 watts and as soon as there is no more PV power wattpilot stops charging the car without draining the home battery.

Next Trip:
- Limit discharge current

The goal is to use this setting if PV power is weak and we want to push energy from the home battery to the car. Because 3 phases are too heavy for a Multiplus 3000 we limit the wattpilot to a single phase. Use wattpilot app to limit the amount of kWh which have to be charged into the car. It will also empty the rest of the home battery if there is some power left.

Default:
- Set Grid Point

The goal is to load the EV with all phases and maximum current without using the battery. This setting is used within the winter because the battery SoC can better be used within the home because of the overall efficiency.

Have fun!